+++ library/init.php
@@ -15,7 +15,7 @@
  * @package     WebMan WordPress Theme Framework
  * @subpackage  Core
  *
- * @version  2.5.4
+ * @version  2.5.5
  *
  * Used global hooks:
  *