+++ changelog.md
@@ -1,5 +1,30 @@
 # Reykjavik Changelog
 
+## 1.0.1
+
+* **Update**: One Click Demo Import demo files moved to the plugin setup folder
+* **Update**: Welcome admin page theme rating link
+* **Update**: Code formatting
+* **Update**: Compatible with WordPress 4.9
+* **Update**: CSS Starter 3.16.0
+* **Update**: Library 2.5.5
+* **Update**: Documentation
+* **Fix**: Author archive page layout
+* **Fix**: Recent Posts widget compatibility with Subtitles plugin
+
+### Files changed:
+
+	changelog.md
+	style.css
+	assets/scss/_setup.scss
+	assets/scss/starter/*.*
+	includes/frontend/class-header.php
+	includes/plugins/one-click-demo-import/class-one-click-demo-import.php
+	includes/widgets/class-wp-widget-recent-posts.php
+	library/*.*
+	templates/parts/admin/welcome-promo.php
+
+
 ## 1.0.0
 
 * Initial release.